+++ cli/helm/diff.go
@@ -1,4 +1,4 @@
-package common
+package helm
 
 import (
 	"sort"
